+++ backend/genesys/gl124.cpp
@@ -1268,8 +1268,8 @@ void CommandSetGl124::init_regs_for_scan(Genesys_Device* dev, const Genesys_Sens
 
   /* y (motor) distance to move to reach scanned area */
   move_dpi = dev->motor.base_ydpi/4;
-    move = static_cast<float>(dev->model->y_offset);
-    move += static_cast<float>(dev->settings.tl_y);
+    move = dev->model->y_offset;
+    move += dev->settings.tl_y;
     move = static_cast<float>((move * move_dpi) / MM_PER_INCH);
   DBG (DBG_info, "%s: move=%f steps\n", __func__, move);
 
@@ -1281,8 +1281,8 @@ void CommandSetGl124::init_regs_for_scan(Genesys_Device* dev, const Genesys_Sens
   DBG(DBG_info, "%s: move=%f steps\n", __func__, move);
 
   /* start */
-    start = static_cast<float>(dev->model->x_offset);
-    start += static_cast<float>(dev->settings.tl_x);
+    start = dev->model->x_offset;
+    start += dev->settings.tl_x;
     start /= sensor.get_ccd_size_divisor_for_dpi(dev->settings.xres);
     start = static_cast<float>((start * sensor.optical_res) / MM_PER_INCH);
