+++ backend/genesys/gl847.cpp
@@ -1041,7 +1041,7 @@ void CommandSetGl847::init_regs_for_scan(Genesys_Device* dev, const Genesys_Sens
 
   move_dpi = dev->motor.base_ydpi;
 
-    move = static_cast<float>(dev->model->y_offset);
+    move = dev->model->y_offset;
     move = static_cast<float>(move + dev->settings.tl_y);
     move = static_cast<float>((move * move_dpi) / MM_PER_INCH);
     move -= dev->head_pos(ScanHeadId::PRIMARY);
@@ -1062,8 +1062,8 @@ void CommandSetGl847::init_regs_for_scan(Genesys_Device* dev, const Genesys_Sens
   DBG(DBG_info, "%s: move=%f steps\n", __func__, move);
 
   /* start */
-    start = static_cast<float>(dev->model->x_offset);
-    start = static_cast<float>(start + dev->settings.tl_x);
+    start = dev->model->x_offset;
+    start = start + dev->settings.tl_x;
     start = static_cast<float>((start * sensor.optical_res) / MM_PER_INCH);
 
     ScanSession session;
@@ -1182,7 +1182,7 @@ SensorExposure CommandSetGl847::led_calibration(Genesys_Device* dev, const Genes
   uint16_t exp[3];
   float move;
 
-    move = static_cast<float>(dev->model->y_offset_calib_white);
+    move = dev->model->y_offset_calib_white;
     move = static_cast<float>((move * (dev->motor.base_ydpi / 4)) / MM_PER_INCH);
     if (move > 20) {
         scanner_move(*dev, dev->model->default_method, static_cast<unsigned>(move),
